pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Олейник Серафим <serafim.oleynik.5777@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pydantic = "^2.9.2"
asyncpg = "^0.30.0"
requests = "^2.32.3"
aiogram = "^3.14.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.10.10"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
